refactor: PostManager 객체 분리 #4
