+++ OpenMarket/OpenMarket/Model/UnboundDataForPost.swift
@@ -1,0 +1,18 @@
+//
+//  UnboundDataForPost.swift
+//  OpenMarket
+//
+//  Created by Jae-hoon Sim on 2022/01/24.
+//
+
+import UIKit
+
+struct UnboundDataForPost {
+    var name: String?
+    var descriptions: String?
+    var price: String?
+    var currency: Currency
+    var discountedPrice: String?
+    var stock: String?
+    var images: [UIImage]?
+}
