+++ OpenMarket/OpenMarket/Network/URLSessionProvider.swift
@@ -15,6 +15,7 @@ class URLSessionProvider {
     init(session: URLSessionProtocol) {
         self.session = session
     }
+
     func request(_ service: OpenMarketService,
                  completionHandler: @escaping (Result<Data, URLSessionProviderError>) -> Void) {
         guard let urlRequest = service.urlRequest else {
